Move to Endorphin.Core.SCPI-based IO
IO is now all dispatched out to Endorphin.Core for enhanced reuseability.  The
instrument is now treated as a base Visa.Instrument rather than wrapped up in a
private type so it can easily be used with the whole SCPI interface.

Function calls are now consistent, with the instrument coming last in all
function calls.